bot/requirements.txt
python-telegram-bot[ext]>=21.0,<22.0
opensearch-py>=2.4,<3.0
python-barcode[images]>=0.15,<1.0
qrcode[pil]>=7.4,<9.0
pyzbar>=0.1.9,<1.0
Pillow>=10.0,<12.0
